BackEnd/requirements.txt
argon2-cffi==19.2.0
bcrypt==3.1.7
cffi==1.14.0
click==7.1.1
Flask==1.1.2
gevent==20.4.0
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
psycopg2-binary==2.8.5
pycparser==2.20
python-dotenv==0.13.0
six==1.14.0
Werkzeug==1.0.1
